chore: rename main content components
