+++ components/NumbersData/NumbersData.tsx
