+++ components/NumbersData/index.ts
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/prefer-default-export
-export { default as DaysList } from "./DaysList";
+export { default as NumbersData } from "./NumbersData";
